pyproject.toml
[tool.poetry]
name = "damn"
version = "0.1.0"
description = "The DAMN (Data Assets Metric Navigation) tool extracts and reports metrics about your data assets"
authors = ["Olivier Dupuis <odupuis@lantrns.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
termcolor = "^2.3.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
damn = "damn.__init__:cli"